debuginfo: extract metadata.rs
